+++ server/logger.cpp
@@ -1,5 +1,6 @@
 #include <cstdio>
 #include <cstdarg>
+#include <time.h>
 
 #include "logger.hpp"
 
@@ -17,6 +18,13 @@ void log_message(LogLevel level, const char * format, ...)
 {
     if (logfile == NULL) return;
     (void)level; // TODO: level ignored
+    char timebuff[64];
+    time_t t;
+    time(&t);
+    struct tm brokendown;
+    localtime_r(&t, &brokendown);
+    strftime(timebuff, sizeof(timebuff), "%Y-%m-%d %H:%M:%S%z", &brokendown);
+    fprintf(logfile, "[%s] ", timebuff);
     va_list args;
     va_start (args, format);
     vfprintf (logfile, format, args);